requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
APScheduler==3.8.1
asgiref==3.4.1
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
Babel==2.12.1
backports.cached-property==1.0.2
beautifulsoup4==4.10.0
boto3==1.18.28
botocore==1.21.28
bs4==0.0.1
cachetools==5.3.1
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.4
chronological==0.1.1
click==8.0.1
clldutils==3.19.0
cmaes==0.9.1
cmake==3.26.4
colorama==0.4.4
colorlog==6.7.0
contourpy==1.1.0
csvw==3.1.3
cycler==0.11.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2
django-apscheduler==0.6.0
django-cors-headers==3.10.1
django-crispy-forms==1.12.0
django-environ==0.4.5
django-formtools==1.0
django-import-export==2.2.0
django-reversion==4.0.0
django-storages==1.11.1
djangorestframework==3.12.4
dlinfo==1.2.1
docopt==0.6.2
einops==0.3.2
et-xmlfile==1.1.0
ffmpeg-python==0.2.0
filelock==3.12.2
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
future==0.15.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.56.2
hmmlearn==0.2.8
httplib2==0.9.2
huggingface-hub==0.16.4
HyperPyYAML==1.2.1
idna==3.2
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==4.8.1
importlib-resources==6.0.0
install==1.3.4
isodate==0.6.1
jmespath==0.10.0
joblib==1.3.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
julius==0.2.7
kiwisolver==1.4.4
language-tags==1.2.0
lazy_loader==0.3
librosa==0.9.2
lit==16.0.6
llvmlite==0.40.1
loguru==0.5.3
lxml==4.6.3
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mysqlclient==2.0.3
networkx==2.8.8
nltk==3.6.3
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
odfpy==1.4.1
omegaconf==2.3.0
openai==0.10.2
openai-whisper==20230314
openpyxl==3.0.7
optuna==3.2.0
packaging==23.1
pandas==1.3.1
pandas-stubs==1.2.0.1
phonemizer==3.2.1
Pillow==10.0.0
pooch==1.6.0
primePy==1.3
proglog==0.1.10
protobuf==3.20.1
psycopg2==2.8.5
pyannote.audio==2.1.1
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.15.1
pylatexenc==2.10
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.0
pytorch-lightning==1.6.5
pytorch-metric-learning==1.7.3
pytz==2021.1
PyYAML==5.4.1
rdflib==6.3.2
referencing==0.29.1
regex==2023.6.3
requests==2.26.0
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3986==1.5.0
rich==13.4.2
rpds-py==0.8.10
rsa==4.9
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
s3transfer==0.5.0
scikit-learn==1.3.0
scipy==1.10.1
segments==2.2.1
semver==2.13.0
sentencepiece==0.1.99
shellingham==1.5.0.post1
simplejson==3.19.1
singledispatchmethod==1.0
six==1.10.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.2.1
soxr==0.3.5
speechbrain==0.5.14
SQLAlchemy==2.0.18
sqlparse==0.4.1
sympy==1.12
tablib==3.0.0
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
threadpoolctl==3.1.0
tiktoken==0.3.1
torch==1.13.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==0.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tqdm==4.62.0
triton==2.0.0
typer==0.9.0
typing_extensions==4.7.1
tzlocal==5.0.1
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.6
Werkzeug==2.3.6
win32-setctime==1.0.3
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
zipp==3.6.0
